+++ src/rules/no-test-only-imports/__test__/project/__test__/h-test.ts
@@ -1,0 +1,3 @@
+import { _testOnlyHelper } from '../h';
+
+console.log(_testOnlyHelper);