+++ src/util/files.ts
@@ -257,8 +257,7 @@ export function isFileIgnored(packageRootDir: string, filePath: string) {
   return false;
 }
 
-// eslint-disable-next-line fast-import/no-unused-exports
-export function _reset() {
+export function _testOnlyResetFiles() {
   ignoreData.clear();
 }